move MD function into MD directory
