+++ kernel/include/dispatch.h
@@ -18,6 +18,7 @@
 #include <barrelfish_kpi/cpu.h>
 #include <barrelfish_kpi/dispatcher_shared_arch.h>
 #include <barrelfish_kpi/dispatcher_shared_target.h>
+#include <dispatcher_arch.h>
 #include <capabilities.h>
 #include <misc.h>
 
@@ -71,14 +72,6 @@ get_dispatcher_shared_generic_cap(struct capability* disp_cap, dispatcher_handle
     dispatcher_handle_t handle = local_phys_to_mem(disp_cap->u.frame.base);
     assert (handle == disp);
     return get_dispatcher_shared_generic(handle);
-}
-
-static inline struct dispatcher_shared_arm*
-get_dispatcher_shared_arm_cap(struct capability* disp_cap, dispatcher_handle_t disp)
-{
-    dispatcher_handle_t handle = local_phys_to_mem(disp_cap->u.frame.base);
-    assert (handle == disp);
-    return get_dispatcher_shared_arm(handle);
 }
 
 static inline bool dispatcher_is_disabled_ip_cap(struct capability* disp_cap,
